Refactor environment code
